Add Databricks Sampler, Refactor Unity Catalog Sampler
