+++ ingestion/src/metadata/ingestion/source/database/unitycatalog/service_spec.py
@@ -11,9 +11,8 @@ from metadata.ingestion.source.database.unitycatalog.usage import (
 from metadata.profiler.interface.sqlalchemy.unity_catalog.profiler_interface import (
     UnityCatalogProfilerInterface,
 )
-from metadata.profiler.interface.sqlalchemy.unity_catalog.sampler_interface import (
-    UnityCatalogSamplerInterface,
-)
+from metadata.sampler.sqlalchemy.unitycatalog.sampler import UnityCatalogSamplerInterface
+
 from metadata.utils.service_spec.default import DefaultDatabaseSpec
 
 ServiceSpec = DefaultDatabaseSpec(
